Added proper description of project and a guide.
Yarp, some parts of the guide might be a bit unpolished but this should do the job if someone wants to try MEAN stack with CRUD and a small angular data connection sample straight off the bat. This ain't the best tutorial but it's a tutorial / starting point. Hopefully someone will find it useful :)